26.5.1
release v26.5.1 — 11 modules formerly biz-gated now ship in the AGPL build

Ports the v26.2.7 module set (firewall, crontab, backup verification,
automated rollback, ssl observatory, docker engine config, wireguard,
image builder, dns providers, calendar, marketplace) plus a proxy-extended
enhancement into the standard AGPL binary. No biz gating, no runtime
caps, no call-home.

11 new migrations (046_calendar through 056_marketplace). Recon (044/045)
untouched. Migration-additive only; no breaking changes.

Full notes: docs/v26.5/release-notes-v26.5.1.md

+++ internal/integrations/git/provider_gitea.go
@@ -601,12 +601,12 @@ func (p *GiteaProvider) CreateRelease(ctx context.Context, repoID string, opts C
 	}
 
 	apiOpts := gitea.CreateReleaseOptions{
-		TagName:    opts.TagName,
-		Name:       opts.Name,
-		Body:       opts.Body,
-		IsDraft:    opts.Draft,
+		TagName:      opts.TagName,
+		Name:         opts.Name,
+		Body:         opts.Body,
+		IsDraft:      opts.Draft,
 		IsPrerelease: opts.Prerelease,
-		Target:     opts.Target,
+		Target:       opts.Target,
 	}
 
 	r, err := p.client.CreateRelease(ctx, owner, name, apiOpts)
@@ -809,7 +809,7 @@ func giteaRepoToModel(r gitea.APIRepository) models.GitRepository {
 		StarsCount:    r.Stars,
 		ForksCount:    r.Forks,
 		OpenIssues:    r.OpenIssues,
-		SizeKB:        int64(r.Size),
+		SizeKB:        r.Size,
 	}
 }
